Move integration test to where it belongs
